requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL only, not needed for SQLite

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Task Queue (Optional for async processing)
# celery==5.3.4  # Optional - not needed for basic setup
# redis==5.0.1  # Optional - not needed for basic setup

# NLP/Text Processing
openai>=1.3.7
# langchain==0.0.340  # Not used in the current implementation
# tiktoken==0.5.1  # Optional - not needed for basic OpenAI usage

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.1  # For TestClient (already present above)

# Logging
loguru==0.7.2

# Web Automation
selenium>=4.15.0
webdriver-manager>=4.0.1
playwright>=1.40.0

# Scheduler
APScheduler>=3.10.4

# System monitoring
psutil>=5.9.6

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0
